reorg(8/8): rename source files for Classic/Home symmetry
Prefixes Classic-side files with `classic-` wherever a matching
`home-` file exists. The rule is: a file gets a `classic-` or
`home-` prefix only when a sibling file for the other side exists.
Files that describe concepts unique to one side (e.g. Classic
`area.ts`, `building.ts`, `floor.ts` — Home only has devices) keep
their unprefixed names.

**Source files renamed:**
- `models/registry.ts`    → `models/classic-registry.ts`
  (sibling: `models/home-registry.ts`)
- `models/device.ts`      → `models/classic-device.ts`
  (sibling: `models/home-device.ts`)
- `facades/manager.ts`    → `facades/classic-manager.ts`
  (sibling: `facades/home-manager.ts`)
- `facades/device-ata.ts` → `facades/classic-device-ata.ts`
  (sibling: `facades/home-device-ata.ts`)

**Test files renamed to mirror source:**
- `tests/unit/melcloud-api.test.ts` → `classic-api.test.ts`
  (sibling: `home-api.test.ts`)
- `tests/unit/registry.test.ts`     → `classic-registry.test.ts`
  (sibling: `home-registry.test.ts`)
- `tests/unit/device.test.ts`       → `classic-device.test.ts`

Files NOT prefixed (no Home counterpart):
- `models/area.ts`, `models/building.ts`, `models/floor.ts`,
  `models/base.ts`, `models/interfaces.ts`, `models/symbols.ts`
- `facades/area.ts`, `facades/building.ts`, `facades/floor.ts`,
  `facades/base*.ts`, `facades/device-atw*.ts`, `facades/device-erv.ts`,
  `facades/factory.ts`, `facades/interfaces.ts`
- `api/interfaces.ts`, `api/sync-manager.ts` (shared infrastructure)
- `types/*` (types are subdivided by device type, not by Classic/Home)

460 tests passing, 100 % coverage preserved.
