+++ src/facades/factory.ts
@@ -4,7 +4,7 @@ import { DeviceType } from '../constants.ts'
 import type { DeviceFacadeAny, Facade } from './interfaces.ts'
 import { AreaFacade } from './area.ts'
 import { BuildingFacade } from './building.ts'
-import { DeviceAtaFacade } from './device-ata.ts'
+import { DeviceAtaFacade } from './classic-device-ata.ts'
 import { DeviceAtwHasZone2Facade } from './device-atw-has-zone2.ts'
 import { DeviceAtwFacade } from './device-atw.ts'
 import { DeviceErvFacade } from './device-erv.ts'
