+++ src/models/classic-device.ts
