+++ tests/unit/classic-registry.test.ts
